+++ quadratic-core/src/controller/operations/formatting.rs
@@ -1,6 +1,9 @@
 use crate::{
     controller::GridController,
-    grid::{formatting::CellFmtArray, NumericCommas, NumericFormat, NumericFormatKind},
+    grid::{
+        formatting::CellFmtArray, generate_borders, BorderSelection, NumericCommas, NumericFormat,
+        NumericFormatKind,
+    },
     RunLengthEncoding, SheetPos, SheetRect,
 };
 
@@ -109,5 +112,57 @@ impl GridController {
                 sheet_rect.len(),
             )),
         }]
+    }
+
+    pub fn clear_formatting_operations(&mut self, sheet_rect: SheetRect) -> Vec<Operation> {
+        let len = sheet_rect.size().len();
+        let mut ops = vec![
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::Align(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::Wrap(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::NumericFormat(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::NumericDecimals(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::Bold(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::Italic(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::TextColor(RunLengthEncoding::repeat(None, len)),
+            },
+            Operation::SetCellFormats {
+                sheet_rect,
+                attr: CellFmtArray::FillColor(RunLengthEncoding::repeat(None, len)),
+            },
+        ];
+
+        // clear borders
+        let sheet = self.grid.sheet_from_id(sheet_rect.sheet_id);
+        let borders = generate_borders(
+            sheet,
+            &sheet_rect.into(),
+            vec![BorderSelection::Clear],
+            None,
+        );
+        ops.push(Operation::SetBorders {
+            sheet_rect,
+            borders,
+        });
+        ops
     }
 }
